Job Deletion Fix
A bug was causing all 'delete' buttons to always delete the most recent
job added.

The deleteJob function was not being passed an argument and was instead
using the last 'job' variable, which was set in an each loop
in the HTML. Fixed the button's on:click property to pass in the
correct job.